Review and refactor deprecated code rules
