+++ tests/linter/rules/deprecated/deprecated_singular_header/test.robot
