+++ tests/linter/rules/deprecated/deprecated_singular_header/test_rule.py
