+++ tests/linter/rules/deprecated/deprecated_statement/expected_output_force_tags.txt
@@ -1,0 +1,1 @@
+force_tags.robot:2:1 [W] DEPR02 'Force Tags' is deprecated since Robot Framework version 6.0, use 'Test Tags' instead
