+++ tests/linter/rules/deprecated/deprecated_statement/templated_suite.robot
