+++ tests/linter/rules/deprecated/deprecated_statement/test.robot
