+++ tests/linter/rules/deprecated/deprecated_statement/test_rule.py
