+++ tests/linter/rules/deprecated/deprecated_with_name/expected_output.txt
@@ -1,0 +1,1 @@
+test.robot:3:22 [W] DEPR03 Deprecated 'WITH NAME' alias marker used instead of 'AS'
