+++ tests/linter/rules/deprecated/deprecated_with_name/test_rule.py
