+++ tests/linter/rules/deprecated/if_can_be_used/test.robot
