+++ tests/linter/rules/deprecated/replace_create_with_var/test_rule.py
