+++ tests/linter/rules/deprecated/replace_set_variable_with_var/test.robot
