--- tests/linter/rules/naming/deprecated_statement/expected_output_force_tags.txt
@@ -1,1 +1,0 @@
-force_tags.robot:2:1 [W] 0319 'Force Tags' is deprecated since Robot Framework version 6.0, use 'Test Tags' instead
