--- tests/linter/rules/naming/deprecated_with_name/expected_output.txt
@@ -1,1 +1,0 @@
-test.robot:3:22 [W] 0321 'WITH NAME' alias marker is deprecated since Robot Framework 6.0 version and will be removed in the future release. Use 'AS' instead
